pyproject.toml
[project]
name = "plant-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.11"
dependencies = [
    "datasets>=4.4.1",
    "evaluate>=0.4.6",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "matplotlib>=3.10.7",
    "mlflow>=3.6.0",
    "nvitop>=1.5.3",
    "pillow>=12.0.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "scipy>=1.11.0",
    "streamlit>=1.51.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "transformers[torch]>=4.57.1",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["plant_classifier*"]
exclude = ["mlruns*", "notebooks*", "scripts*"]
